+++ services/iaas/src/stackit/iaas/models/create_server_payload.py
@@ -25,6 +25,7 @@ from pydantic import (
     BaseModel,
     ConfigDict,
     Field,
+    StrictBool,
     StrictBytes,
     StrictStr,
     field_validator,
@@ -55,6 +56,9 @@ class CreateServerPayload(BaseModel):
         alias="availabilityZone",
     )
     boot_volume: Optional[BootVolume] = Field(default=None, alias="bootVolume")
+    config_drive: Optional[StrictBool] = Field(
+        default=False, description="When true the server is created with a config drive.", alias="configDrive"
+    )
     created_at: Optional[datetime] = Field(
         default=None, description="Date-time when resource was created.", alias="createdAt"
     )
@@ -118,6 +122,7 @@ class CreateServerPayload(BaseModel):
         "agent",
         "availabilityZone",
         "bootVolume",
+        "configDrive",
         "createdAt",
         "errorMessage",
         "id",
@@ -338,6 +343,7 @@ class CreateServerPayload(BaseModel):
                 "agent": ServerAgent.from_dict(obj["agent"]) if obj.get("agent") is not None else None,
                 "availabilityZone": obj.get("availabilityZone"),
                 "bootVolume": BootVolume.from_dict(obj["bootVolume"]) if obj.get("bootVolume") is not None else None,
+                "configDrive": obj.get("configDrive") if obj.get("configDrive") is not None else False,
                 "createdAt": obj.get("createdAt"),
                 "errorMessage": obj.get("errorMessage"),
                 "id": obj.get("id"),